requirements.txt
absl-py==1.1.0
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Authlib==0.15.5
autopep8==1.6.0
backcall==0.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
dataclasses-json==0.5.7
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
discord.py==2.1.0
distro-info==1.0
Django==4.1
entrypoints==0.4
executing==0.8.3
fonttools==4.33.3
frozenlist==1.3.1
gyp==0.1
idna==3.3
ipykernel==6.13.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.2
marshmallow==3.16.0
marshmallow-enum==1.5.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mediapipe==0.8.10
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.19.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pycurl==7.43.0.6
Pygments==2.7.1
PyGObject==3.38.0
pyparsing==3.0.9
python-apt==2.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-twitter-v2==0.7.7
pytz==2022.1
PyYAML==5.3.1
pyzmq==23.1.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.2.2.post1
tweepy==4.10.0
typing-extensions==4.2.0
typing-inspect==0.7.1
unattended-upgrades==0.1
urllib3==1.26.9
wcwidth==0.2.5
yarl==1.8.1
